Add files via upload
Anomalies are pattern in the data that do not conform to a well defined normal behavior. The cause of anomaly may be a
malicious activity or some kind of intrusion. This abnormal behavior found in the dataset is interesting to the analyst and this is
the most important feature for anomaly detection